Move readDir methods to a utils file
